% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffabs_elements.R
\name{diffabs_elements}
\alias{diffabs_elements}
\title{Converts spectra to element count values}
\usage{
diffabs_elements(spectra, rois)
}
\arguments{
\item{spectra}{the output of diffabs_spectra()}

\item{rois}{nested list of regions of interest along the spectral
domain}
}
\value{

}
\description{
Uses the output of `diffabs_spectra()` to
calculate element counts based upon regions
of interest list. Regions of interest are
defined by the elements of interest and
beamline / detector configurations.
}
\details{
Please consult with DIFFABS engineers for
setting details.
}
