% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffabs_spectra.R
\name{diffabs_spectra}
\alias{diffabs_spectra}
\title{convert DIFFABS output to corrected spectra}
\usage{
diffabs_spectra(file, channels)
}
\arguments{
\item{file}{an NXS DIFFABS file}

\item{channels}{a list of detector channels to use}
}
\value{

}
\description{
Reads in a NXS file and returns spectra, making
the necessary corrections
}
