% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffabs_tidy.R
\name{diffabs_tidy}
\alias{diffabs_tidy}
\title{Tidy the output of an elemental analysis}
\usage{
diffabs_tidy(spectra)
}
\arguments{
\item{spectra}{the output of `diffabs_elemental_analysis()`}
}
\value{

}
\description{
Converts nested list output of `diffabs_elemental_analysis()` to a
tidy format for easy processing.
}
